requirements.txt
langchain
langchain_community
langchain_openai
langgraph
langgraph_supervisor
tavily-python
dotenv
bs4
faiss-cpu
pypdf
duckduckgo-search
pycricbuzz
requests_html
lxml_html_clean
cloudscraper
undetected_chromedriver
distutils